backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
requests>=2.28.0
sqlalchemy>=2
pyyaml>=6.0.0
aiofiles>=23.1.0 